+++ src/clj/rems/css/styles.clj
@@ -321,7 +321,7 @@
    [:a
     :button
     {:cursor :pointer
-     :color (get-theme-attribute :link-color "#007bff")}
+     :color (get-theme-attribute :link-color "#025b96")}
     [:&:hover {:color (get-theme-attribute :link-hover-color :color4)}]]
    [:.pointer {:cursor :pointer}
     [:label.form-check-label {:cursor :pointer}]]
@@ -643,9 +643,9 @@
    [:#action-commands {:display "flex"
                        :flex-flow "row wrap"
                        :margin-bottom (u/em -0.5)}
-    [(s/> "*")
+    ["> *"
      {:margin-bottom (u/em 0.5)}]
-    [(s/> "*:not(:last-child)")
+    ["> *:not(:last-child)"
      {:margin-right (u/em 0.5)}]]
 
    [:.event-comment {:white-space :pre-wrap}]