+++ src/cljs/rems/actions/request_decision.cljs
@@ -1,6 +1,6 @@
 (ns rems.actions.request-decision
   (:require [re-frame.core :as rf]
-            [rems.actions.components :refer [action-attachment action-button comment-field action-form-view button-wrapper command! user-selection]]
+            [rems.actions.components :refer [action-attachment action-link comment-field action-form-view button-wrapper command! user-selection]]
             [rems.text :refer [text]]))
 
 (def ^:private action-form-id "request-decision")
@@ -26,27 +26,27 @@
               :on-finished on-finished})
    {}))
 
-(defn request-decision-action-button []
-  [action-button {:id action-form-id
-                  :text (text :t.actions/request-decision)
-                  :on-click #(rf/dispatch [::open-form])}])
+(defn request-decision-action-link []
+  [action-link {:id action-form-id
+                :text (text :t.actions/request-decision-dropdown-from-user)
+                :on-click #(rf/dispatch [::open-form])}])
 
 (defn request-decision-view
   [{:keys [application-id disabled on-send]}]
   [action-form-view action-form-id
-   (text :t.actions/request-decision)
+   (text :t.actions/request-decision-from-user)
    [[button-wrapper {:id "request-decision"
                      :text (text :t.actions/request-decision)
                      :class "btn-primary"
                      :on-click on-send
                      :disabled disabled}]]
    [:div
+    [user-selection {:field-key action-form-id
+                     :subscription [:rems.actions.components/deciders]}]
     [comment-field {:field-key action-form-id
                     :label (text :t.form/add-comments-not-shown-to-applicant)}]
     [action-attachment {:field-key action-form-id
-                        :application-id application-id}]
-    [user-selection {:field-key action-form-id
-                     :subscription [:rems.actions.components/deciders]}]]])
+                        :application-id application-id}]]])
 
 (defn request-decision-form [application-id on-finished]
   (let [selected-deciders @(rf/subscribe [:rems.actions.components/users action-form-id])